feat: daemon merge convert_config as a subcommand, Python config convert in-memory
