+++ rnsd-rs/src/config.rs
@@ -1,5 +1,9 @@
-use serde::{Deserialize, Serialize};
+use std::ffi::OsStr;
+use std::fs;
 use std::path::{Path, PathBuf};
+
+use regex::Regex;
+use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize, Default)]
 pub struct Config {
@@ -137,6 +141,142 @@ fn default_shared_port() -> u16 { 37428 }
 fn default_control_port() -> u16 { 37429 }
 fn default_loglevel() -> u8 { 4 }
 
+pub fn migrate_config(config_file: &Path) -> Result<(), Box<dyn std::error::Error>> {
+    if !config_file.exists() {
+        eprintln!("Error: File '{}' does not exist", config_file.display());
+        std::process::exit(1);
+    }
+    println!("Reading config from: {}", config_file.display());
+    let content = fs::read_to_string(config_file)?;
+    if toml::from_str::<Config>(&content).is_ok() {
+        println!("File is already a valid TOML config: exiting");
+        return Ok(())
+    }
+    let converted = convert_config(&content);
+    // validate
+    match toml::from_str::<toml::Value>(&converted) {
+        Ok(_) => {}
+        Err(err) => {
+            eprintln!("error: converted text is not a valid TOML file");
+            return Err(err.into())
+        }
+    }
+    if cfg!(debug_assertions) {
+        match toml::from_str::<Config>(&converted) {
+            Ok(_) => {}
+            Err(err) => {
+                eprintln!("error: converted text is not a valid rnsd-rs Config file");
+                return Err(err.into())
+            }
+        }
+    }
+    // in case the passed-in file already has a .toml extension, create a backup to prevent
+    // overwriting it
+    let new_config_file = if config_file.extension() == Some(OsStr::new("toml")) {
+        let backup_path = config_file.with_extension("bak");
+        fs::write(&backup_path, &content)?;
+        println!("Created backup at: {}", backup_path.display());
+        config_file.to_owned()
+    } else {
+        config_file.with_extension("toml")
+    };
+    fs::write(&new_config_file, &converted)?;
+    println!("✓ Converted config written to: {}", new_config_file.display());
+    println!();
+    println!("Changes made:");
+    println!("  - Converted True/False/Yes/No → true/false");
+    println!("  - Quoted all string values (IPs, hostnames, paths, types)");
+    println!("  - Converted [[Interface Name]] → [[interfaces]] with name field");
+    println!("  - Normalized indentation");
+    println!("  - Preserved all comments");
+    Ok(())
+}
+
+fn convert_config(content: &str) -> String {
+    fn quote_if_needed(line: &str, key: &str) -> String {
+        let pattern = format!("{} = ", key);
+        let quoted_pattern = format!("{} = \"", key);
+        // Already quoted or not present
+        if !line.contains(&pattern) || line.contains(&quoted_pattern) {
+            return line.to_string();
+        }
+        // Find the value
+        if let Some(pos) = line.find(&pattern) {
+            let value_start = pos + pattern.len();
+            let rest = &line[value_start..];
+            let value = rest.split_whitespace().next().unwrap_or(rest).trim();
+            // Don't quote numbers or booleans
+            if value.parse::<i64>().is_ok() || value.parse::<f64>().is_ok() 
+                || value == "true" || value == "false" {
+                return line.to_string();
+            }
+            // Quote the value
+            format!("{}{} = \"{}\"", &line[..pos], key, value)
+        } else {
+            line.to_string()
+        }
+    }
+
+    let mut output = String::new();
+    let re_false = Regex::new(r" = \b(No|no|False)\b").unwrap();
+    let re_true = Regex::new(r" = \b(Yes|yes|True)\b").unwrap();
+    let re_nil = Regex::new(r"^(\w+)\s*=\s*\b(None|none|nil|Nil|null|Null)\b").unwrap();
+    for line in content.lines() {
+        let trimmed = line.trim();
+        // Empty lines pass through
+        if trimmed.is_empty() {
+            output.push('\n');
+            continue;
+        }
+        // Skip [interfaces] header - we use [[interfaces]] instead
+        if trimmed == "[interfaces]" {
+            continue;
+        }
+        // Detect interface block start
+        if trimmed.starts_with("[[") && trimmed.ends_with("]]") {
+            let name = trimmed.trim_start_matches("[[").trim_end_matches("]]").trim();
+            if name != "interfaces" {
+                // Convert [[Interface Name]] to [[interfaces]]
+                output.push_str("\n[[interfaces]]\n");
+                output.push_str(&format!("name = \"{}\"\n", name));
+                continue;
+            } else {
+                output.push_str("\n[[interfaces]]\n");
+                continue;
+            }
+        }
+        // Process the line
+        let mut converted = trimmed.to_string();
+        // Convert booleans
+        converted = re_false.replace_all(&converted, " = false").to_string();
+        converted = re_true.replace_all(&converted, " = true").to_string();
+        // Comment out nil values, as toml does not support them (https://github.com/toml-lang/toml/issues/30)
+        if re_nil.is_match(&converted) {
+            converted = format!("# {}", converted);
+            output.push_str(&converted);
+            output.push('\n');
+            continue;
+        }
+        // Quote unquoted string values (only for non-comments)
+        if !converted.starts_with('#') {
+            converted = quote_if_needed(&converted, "type");
+            converted = quote_if_needed(&converted, "remote");
+            converted = quote_if_needed(&converted, "target_host");
+            converted = quote_if_needed(&converted, "bind_host");
+            converted = quote_if_needed(&converted, "listen_ip");
+            converted = quote_if_needed(&converted, "forward_ip");
+            converted = quote_if_needed(&converted, "peers");
+            converted = quote_if_needed(&converted, "instance_name");
+            converted = quote_if_needed(&converted, "port");
+            converted = quote_if_needed(&converted, "callsign");
+            converted = quote_if_needed(&converted, "parity");
+        }
+        output.push_str(&converted);
+        output.push('\n');
+    }
+    output
+}
+
 impl Default for ReticulumConfig {
     fn default() -> Self {
         Self {
@@ -180,27 +320,38 @@ impl Config {
     }
 
     pub fn from_file(path: &Path) -> Result<Self, Box<dyn std::error::Error>> {
-        let config_file = if path.join("config").exists() {
-            path.join("config")
+        let config_basename = if path.join("config.toml").exists() {
+            "config.toml"
+        } else if path.join("config").exists() {
+            "config"
         } else {
-            path.join("config.toml")
+            let err = format!("no config.toml or config file found in config path {}",
+                path.display());
+            return Err(err.into())
         };
-        let content = std::fs::read_to_string(&config_file)?;
-        let config: Self = toml::from_str(&content).inspect_err(|_e| {
-            eprintln!();
-            eprintln!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-            eprintln!("INVALID TOML FORMAT");
-            eprintln!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-            eprintln!();
-            eprintln!("Your config file appears to be in Python Reticulum format.");
-            eprintln!("Use the converter tool to migrate it to standard TOML:");
-            eprintln!();
-            eprintln!("  cargo run --example convert_config -- {}", config_file.display());
-            eprintln!();
-            eprintln!("This will create a backup and convert your config to valid TOML.");
-            eprintln!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-            eprintln!();
-        })?;
+        let config_file = path.join(config_basename);
+        let content = fs::read_to_string(&config_file)?;
+        let config: Self = match toml::from_str(&content) {
+            Ok(config) => config,
+            Err(err) => {
+                if config_basename == "config.toml" {
+                    eprintln!("{config_file:?} is not valid TOML");
+                    return Err(err.into())
+                } else {
+                    // attempt to convert
+                    eprintln!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
+                    eprintln!("Your config file appears to be in Python Reticulum format.");
+                    eprintln!("You can use the converter tool to migrate it to standard TOML:");
+                    eprintln!();
+                    eprintln!("  cargo run -p rnsd-rs -- convert-config {}", config_file.display());
+                    eprintln!();
+                    eprintln!("This command will create a backup and convert your config to valid TOML.");
+                    eprintln!("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
+                    let converted = convert_config(&content);
+                    toml::from_str(&converted)?
+                }
+            }
+        };
         if config.reticulum.share_instance {
             log::warn!("share_instance is enabled but shared instances are not supported in reticulum-rs");
             log::warn!("Each Rust daemon process runs independently and is only limited by available ports");
@@ -219,10 +370,10 @@ impl Config {
         } else {
             log::warn!("No existing configuration found, creating default config");
             let default_dir = Self::default_path();
-            std::fs::create_dir_all(&default_dir)?;
+            fs::create_dir_all(&default_dir)?;
             let config = Self::default_config();
             let config_file = default_dir.join("config.toml");
-            std::fs::write(&config_file, toml::to_string_pretty(&config)?)?;
+            fs::write(&config_file, toml::to_string_pretty(&config)?)?;
             log::warn!("Created default configuration at: {}", config_file.display());
             log::warn!("Please review and customize the configuration for your needs");
             Ok((config, default_dir))
